+++ src/server/modules/user/interfaces/ApiKeyAuthServiceInterface.ts
@@ -1,6 +1,5 @@
 interface ApiKeyAuthServiceInterface {
   createApiKey: (userId: number) => Promise<string>
-  getApiKeyHash: (apiKey: string) => Promise<string>
 }
 
 export default ApiKeyAuthServiceInterface
